+++ vcd/resource_vcd_vapp_access_control_test.go
@@ -333,94 +333,3 @@ resource "vcd_vapp_access_control" "{{.AccessControlIdentifier3}}" {
   }
 }
 `
-
-func TestAccVcdVappAccessControlResourceNotFound(t *testing.T) {
-	preTestChecks(t)
-	skipTestForApiToken(t)
-	// This test invokes go-vcloud-director SDK directly
-	if vcdShortTest {
-		t.Skip(acceptanceTestsSkipped)
-		return
-	}
-
-	var params = StringMap{
-		"Org":                 testConfig.VCD.Org,
-		"Vdc":                 testConfig.Nsxt.Vdc,
-		"SharedToEveryone":    "true",
-		"EveryoneAccessLevel": fmt.Sprintf(`everyone_access_level = "%s"`, types.ControlAccessReadWrite),
-		"VappName":            t.Name(),
-		"Tags":                "vapp",
-	}
-	testParamsNotEmpty(t, params)
-
-	configText := templateFill(testAccVappAccessControlResourceNotFound, params)
-
-	debugPrintf("#[DEBUG] CREATION CONFIGURATION: %s", configText)
-
-	cachedId := &testCachedFieldValue{}
-
-	resourceName := "vcd_vapp_access_control.ac1"
-	resource.Test(t, resource.TestCase{
-		ProviderFactories: testAccProviders,
-		CheckDestroy:      testAccCheckVappAccessControlDestroy(testConfig.VCD.Org, testConfig.Nsxt.Vdc, []string{params["VappName"].(string)}),
-		Steps: []resource.TestStep{
-			{
-				Config: configText,
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckVcdVappAccessControlExists(resourceName, testConfig.VCD.Org, testConfig.Nsxt.Vdc),
-					cachedId.cacheTestResourceFieldValue(resourceName, "id"),
-				),
-			},
-			{
-				// This function finds newly created resource and deletes it before
-				// next plan check
-				PreConfig: func() {
-					vcdClient := createSystemTemporaryVCDConnection()
-					org, err := vcdClient.GetAdminOrgByName(params["Org"].(string))
-					if err != nil {
-						t.Errorf("error: could not find Org: %s", err)
-					}
-					vdc, err := org.GetVDCByName(params["Vdc"].(string), false)
-					if err != nil {
-						t.Errorf("error: could not find VDC: %s", err)
-					}
-
-					vapp, err := vdc.GetVAppById(cachedId.fieldValue, false)
-					if err != nil {
-						t.Errorf("could not find vApp %s: %s", cachedId.fieldValue, err)
-					}
-					task, err := vapp.Delete()
-					if err != nil {
-						t.Errorf("could not remove vApp '%s': %s", cachedId.fieldValue, err)
-					}
-
-					err = task.WaitTaskCompletion()
-					if err != nil {
-						t.Errorf("vApp removal task is failing '%s': %s", cachedId.fieldValue, err)
-					}
-				},
-				// Expecting to get a non-empty plan because resource was removed using SDK in
-				// PreConfig
-				Config:             configText,
-				PlanOnly:           true,
-				ExpectNonEmptyPlan: true,
-			},
-		},
-	})
-	postTestChecks(t)
-}
-
-const testAccVappAccessControlResourceNotFound = `
-resource "vcd_vapp" "test1" {
-  name = "{{.VappName}}"
-}
-
-resource "vcd_vapp_access_control" "ac1" {
-  org      = "{{.Org}}"
-  vdc      = "{{.Vdc}}"
-  vapp_id  = vcd_vapp.test1.id
-
-  shared_with_everyone    = {{.SharedToEveryone}}
-  {{.EveryoneAccessLevel}}
-}
-`
